controller: remove all the old code and move everything to the new code.
This implies removing aiostem.reply and aiostem.event, as well as
argument.py, message.py, fix the controller code and then the tests.

This also removes the dependency on aiofiles and stem as they are
no longer needed. Some parts of the old code will return in another
form when crypto will be dealt with, in a coming future.

Tests around the controller were also cleaned up and enhanced so we
keep a good confidence in the present code.

Related parts of the documentation were also removed, more is still
to be written on this topic, this will be done in a second time.

Signed-off-by: Romain Bezut <morian@xdec.net>