syntax: do not allow optional arguments along with keyword arguments.
This was found because `CIRC_MINOR` was generating backtraces due to an
optional argument not being set. As a result the first keyword argument
was mistaken for the missing optional argument and all went wrong since
then.

This patch adds an additional runtime check on our ReplySyntax class
so that this situation can be quickly found out during development.
Additionally, several events and replies have been fixed for this
specific issue, including `EventCircMinor`.

The fix is simply to avoid optional argument but consider any keyword
argument without a key part as part of the map. It only works when a
single optional argument is possible, but luckily this is always true.

Signed-off-by: Romain Bezut <morian@xdec.net>