structures: caches on HsDescV3 are now bound to instances.
`functools.cache` is a simple wrapper around `functools.lru_cache`, and
a naive use means that the whole descriptor instance is kept in cache.
More on this on: https://rednafi.com/python/lru_cache_on_methods/

This patch binds the cache to the descriptor using a `__post_init__`
method, so that descriptors can be garbage collected appropriately.

Signed-off-by: Romain Bezut <morian@xdec.net>